+++ backend/src/main/java/model/PagoOrden.java
@@ -1,6 +1,7 @@
 package model;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record PagoOrden(String id, double total, List<Orden> orden) {
+public record PagoOrden(String id, BigDecimal total, List<Orden> orden) {
 }
