+++ backend/src/main/java/model/PagoTotal.java
@@ -1,3 +1,5 @@
 package model;
 
-public record PagoTotal(String id, double total, Cuenta cuenta) { }
+import java.math.BigDecimal;
+
+public record PagoTotal(String id, BigDecimal total, Cuenta cuenta) { }
