+++ backend/src/main/java/model/Plato.java
@@ -1,3 +1,5 @@
 package model;
 
-public record Plato(String id, String nombre, Categoria categoria, String descripcion, double price, Boolean activo) {}
+import java.math.BigDecimal;
+
+public record Plato(String id, String nombre, Categoria categoria, String descripcion, BigDecimal price, Boolean activo) {}
